5159 - service\databaseService: Pass tables from outside
