--- src/meta/metadata/export.ts
@@ -1,7 +1,0 @@
-export type ExportedTableData<T = unknown> = {
-  schema: string
-  table: string
-  rows: Array<T>
-  exportedAt: string
-  rowCount: number
-}
