+++ src/server/service/databaseService/import.ts
@@ -1,11 +1,13 @@
 import * as pgPromise from 'pg-promise'
 import { Promises } from 'utils/promises'
 
-import { ExportedTableData } from 'meta/metadata/export'
-
 import { BaseProtocol, DB } from 'server/db/db'
 
-const importTable = async (exportedTableData: ExportedTableData, client: BaseProtocol = DB): Promise<void> => {
+import { ExportedTableData } from './export'
+
+type ImportTableProps = { exportedTableData: ExportedTableData }
+const importTable = async (props: ImportTableProps, client: BaseProtocol = DB): Promise<void> => {
+  const { exportedTableData } = props
   const { rows, schema, table } = exportedTableData
 
   if (rows.length === 0) {
@@ -33,8 +35,13 @@ const importTable = async (exportedTableData: ExportedTableData, client: BasePro
   await client.none(query)
 }
 
-export const importTables = async (allData: Array<ExportedTableData>, client: BaseProtocol = DB): Promise<void> => {
-  await Promises.each(allData, async (exportedTableData) => {
-    await importTable(exportedTableData, client)
+type Props = {
+  tables: Array<ExportedTableData>
+}
+
+export const importTables = async (props: Props, client: BaseProtocol = DB): Promise<void> => {
+  const { tables } = props
+  await Promises.each(tables, async (exportedTableData) => {
+    await importTable({ exportedTableData }, client)
   })
 }
