+++ src/tools/data/export.ts
@@ -8,12 +8,14 @@ import { DatabaseService } from 'server/service/databaseService'
 
 import { _copyStaticFixtures } from './_copyStaticFixtures'
 import { _writeTableToFile } from './_writeTableToFile'
+import { EXPORT_TABLES } from './EXPORT_TABLES'
 
 const OUTPUT_DIR = path.join(__dirname, 'fixtures')
 
 const exec = async (): Promise<void> => {
   await DB.tx(async (t) => {
-    const allData = await DatabaseService.exportTables(t)
+    const tables = EXPORT_TABLES.filter((t) => !t.skipExport)
+    const allData = await DatabaseService.exportTables({ tables }, t)
 
     allData.forEach((tableData) => {
       _writeTableToFile(tableData, OUTPUT_DIR)
