+++ src/tools/data/import.ts
@@ -11,13 +11,9 @@ import { DB } from 'server/db/db'
 import { AssessmentRepository } from 'server/db/repository/assessment/assessment'
 import { Schemas } from 'server/db/schemas'
 import { DatabaseService } from 'server/service/databaseService'
-import {
-  EXPORT_ASSESSMENTS,
-  EXPORT_ASSESSMENTS_CYCLES,
-  EXPORT_TABLES,
-} from 'server/service/databaseService/EXPORT_TABLES'
 
 import { _readTableFromFile } from './_readTableFromFile'
+import { EXPORT_ASSESSMENTS, EXPORT_ASSESSMENTS_CYCLES, EXPORT_TABLES } from './EXPORT_TABLES'
 
 const INPUT_DIR = path.join(__dirname, 'fixtures')
 
@@ -47,12 +43,12 @@ const exec = async (): Promise<void> => {
 
   await DB.tx(async (t) => {
     // Use EXPORT_TABLES order to respect foreign key dependencies
-    const allData = EXPORT_TABLES.map((tableConfig) => {
+    const tables = EXPORT_TABLES.map((tableConfig) => {
       const { schema, table } = tableConfig
       return _readTableFromFile(schema, table, INPUT_DIR)
     })
 
-    await DatabaseService.importTables(allData, t)
+    await DatabaseService.importTables({ tables }, t)
 
     // create views
     const assessments = await AssessmentController.getAll({}, t)
